% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R, R/load_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download a Behaverse dataset}
\usage{
download_dataset(name, dest = NULL)

download_dataset(name, dest = NULL)
}
\arguments{
\item{name}{The name of the dataset to download}

\item{dest}{The destination directory to save the dataset}
}
\description{
Download a Behaverse dataset

Download a Behaverse dataset
}
